chore: item component moved to components folder
