+++ pages/profile/sold.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import Layout from "../../components/layout";
-import Item from "../community/item";
+import Item from "../../components/item";
 
 const Sold: NextPage = () => {
   return (
